script command caller moved to its own class and added unit test for it
