+++ script-wrapper/src/main/groovy/org/cedar/psi/wrapper/stream/ScriptWrapperStreamConfig.groovy
@@ -62,32 +62,7 @@ class ScriptWrapperStreamConfig {
     def builder = new StreamsBuilder()
     KStream inputStream = builder.stream(inputTopic)
     inputStream.mapValues ({ msg ->
-      List<String> commandList = command.split(' ') .toList()
-      log.info "Running : $commandList"
-      try {
-        def stdout = new StringBuilder()
-        def stderr = new StringBuilder()
-        def cmd = commandList.execute()
-        cmd.consumeProcessOutput(stdout, stderr)
-        cmd.withOutputStream { OutputStream os ->
-          os << msg
-        }
-        cmd.waitForOrKill(timeout)
-        if (cmd.exitValue()) {
-          log.error "Process exited with non-zero exit code"
-          log.error "Process stdout: ${stdout}"
-          log.error "Process stderr: ${stderr}"
-          return 'ERROR: ' + stderr
-        }
-        else {
-          log.info "Publishing to $outputTopic: $stdout"
-          return stdout.toString()
-        }
-      }
-      catch(Exception e){
-        log.error("Caught exception $e: ${e.message}")
-        return 'ERROR: ' + e.message
-      }
+      ScriptWrapperFunctions.scriptCaller(msg, command, timeout)
     })
     .filterNot({key, msg -> msg.toString().startsWith('ERROR')})
     .mapValues({msg -> doIsoConversion ? isoConversionUtil.parseXMLMetadata(msg as String)  : msg})
